latest

+++ workspace/linux-support/src/express_data_path/c/XDP_PACKET_HEADROOM.rs
@@ -2,4 +2,4 @@
 // Copyright © 2020 The developers of linux-support. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/linux-support/master/COPYRIGHT.
 
 
-pub(crate) const XDP_PACKET_HEADROOM: u32 = 256;
+pub(crate) const XDP_PACKET_HEADROOM: usize = 256;
